+++ ton_client/src/debot/activity.rs
@@ -1,5 +1,5 @@
 
-/// [UNSTABLE](UNSTABLE.md) Describes how much funds will be debited from the target
+/// [UNSTABLE](UNSTABLE.md) [DEPRECATED](DEPRECATED.md) Describes how much funds will be debited from the target
 ///  contract balance as a result of the transaction.
 #[derive(Serialize, Deserialize, Clone, Debug, ApiType, PartialEq)]
 pub struct Spending {
@@ -9,7 +9,7 @@ pub struct Spending {
     pub dst: String,
 }
 
-/// [UNSTABLE](UNSTABLE.md) Describes the operation that the DeBot wants to perform.
+/// [UNSTABLE](UNSTABLE.md) [DEPRECATED](DEPRECATED.md) Describes the operation that the DeBot wants to perform.
 #[derive(Serialize, Deserialize, Clone, Debug, ApiType)]
 #[serde(tag="type")]
 pub enum DebotActivity {