+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 
 
+## [1.44.1] – 2023-07-25
+
+### New
+
+- functions with callbacks (e.g. `processing.process_messages`) can be called as sync.
+- `send_event` parameter is now optional with default value `false`.
+  
+### Deprecated
+- Debot module is [DEPRECATED](./docs/reference/types-and-methods/DEPRECATED.md)
+
 ## [1.44.0] – 2023-07-12
 
 ### New